+++ tests/test/react/entity_reactions.rs
@@ -72,28 +72,28 @@ fn on_any_entity_mutation(In(entity): In<Entity>, mut c: Commands) -> RevokeToke
             mut recorder: ResMut<TestReactRecorder>
         |
         {
-            if let Some(_) = insertion.try_read()
+            if let Some(_) = insertion.get()
             {
                 recorder.0 += 1;
                 assert!(mutation.is_empty());
                 assert!(removal.is_empty());
                 assert!(despawn.is_empty());
             }
-            if let Some(_) = mutation.try_read()
+            if let Some(_) = mutation.get()
             {
                 recorder.0 += 10;
                 assert!(insertion.is_empty());
                 assert!(removal.is_empty());
                 assert!(despawn.is_empty());
             }
-            if let Some(_) = removal.try_read()
+            if let Some(_) = removal.get()
             {
                 recorder.0 += 100;
                 assert!(insertion.is_empty());
                 assert!(mutation.is_empty());
                 assert!(despawn.is_empty());
             }
-            if let Some(_) = despawn.try_read()
+            if let Some(_) = despawn.get()
             {
                 recorder.0 += 1000;
                 assert!(insertion.is_empty());
@@ -116,7 +116,7 @@ fn on_mutation_recursive(mut c: Commands) -> RevokeToken
             mut recorder      : ResMut<TestReactRecorder>
         |
         {
-            let entity = match (insertion.try_read(), mutation.try_read())
+            let entity = match (insertion.get(), mutation.get())
             {
                 (Some(entity), None) => entity,
                 (None, Some(entity)) => entity,
@@ -147,7 +147,7 @@ fn register_reader_for_insertion_event(In(entity): In<Entity>, mut c: Commands) 
                 mut recorder: ResMut<TestReactRecorder>
             |
             {
-                assert_eq!(insertion.read(), entity);
+                assert_eq!(insertion.entity(), entity);
                 assert!(mutation.is_empty());
                 assert!(removal.is_empty());
                 assert!(despawn.is_empty());
@@ -169,7 +169,7 @@ fn register_reader_for_mutation_event(In(entity): In<Entity>, mut c: Commands) -
             |
             {
                 assert!(insertion.is_empty());
-                assert_eq!(mutation.read(), entity);
+                assert_eq!(mutation.entity(), entity);
                 assert!(removal.is_empty());
                 assert!(despawn.is_empty());
                 recorder.0 = 10;
@@ -191,7 +191,7 @@ fn register_reader_for_removal_event(In(entity): In<Entity>, mut c: Commands) ->
             {
                 assert!(insertion.is_empty());
                 assert!(mutation.is_empty());
-                assert_eq!(removal.read(), entity);
+                assert_eq!(removal.entity(), entity);
                 assert!(despawn.is_empty());
                 recorder.0 = 100;
             }
@@ -213,7 +213,7 @@ fn register_reader_for_despawn_event(In(entity): In<Entity>, mut c: Commands) ->
                 assert!(insertion.is_empty());
                 assert!(mutation.is_empty());
                 assert!(removal.is_empty());
-                assert_eq!(despawn.read(), entity);
+                assert_eq!(despawn.entity(), entity);
                 recorder.0 = 1000;
             }
         )